feat: Join CQRS Commands/Queries with handlers, add FluentValidation
