--- TestApp/CQRS/Commands/CommandEntities/DeleteHeroCommand.cs
@@ -1,12 +1,0 @@
-﻿namespace TestApp.CQRS.Commands.CommandEntities
-{
-    public class DeleteHeroCommand : IRequest<bool>
-    {
-        public int Id { get; }
-
-        public DeleteHeroCommand(int id)
-        {
-            Id = id;
-        }
-    }
-}
