--- TestApp/CQRS/Queries/QueryEntities/FindHeroByIdQuery.cs
@@ -1,12 +1,0 @@
-﻿namespace TestApp.CQRS.Queries.QueryEntities
-{
-    public class FindHeroByIdQuery : IRequest<SuperHero>
-    {
-        public int Id { get; }
-
-        public FindHeroByIdQuery(int id)
-        {
-            Id = id;
-        }
-    }
-}
