--- TestApp/CQRS/Queries/QueryEntities/GetAllHeroesQuery.cs
@@ -1,6 +1,0 @@
-﻿namespace TestApp.CQRS.Queries.QueryEntities
-{
-    public class GetAllHeroesQuery : IRequest<List<SuperHero>>
-    {
-    }
-}
