+++ TestApp/Core/GlobalUsingStatements.cs
@@ -2,6 +2,7 @@
 global using TestApp.Models;
 global using Microsoft.EntityFrameworkCore;
 global using TestApp.Data;
-global using TestApp.CQRS.Queries.QueryEntities;
-global using TestApp.CQRS.Commands.CommandEntities;
+global using TestApp.CQRS.Queries;
+global using TestApp.CQRS.Commands;
 global using Serilog;
+global using FluentValidation;
