+++ TestApp/Models/AddHeroModel.cs
@@ -1,0 +1,6 @@
+﻿namespace TestApp.Models
+{
+    public class AddHeroModel : SuperHero
+    {
+    }
+}
