+++ TestApp/Validators/AddHeroModelValidation.cs
@@ -1,0 +1,14 @@
+﻿namespace TestApp.Validators
+{
+    public class AddHeroModelValidation : AbstractValidator<AddHeroModel>
+    {
+        public AddHeroModelValidation()
+        {
+            RuleFor(x => x).NotEmpty();
+            RuleFor(x => x.Id).Equal(0);
+            RuleFor(x => x.Name).NotEmpty();
+            RuleFor(x => x.FirstName).NotEmpty();
+            RuleFor(x => x.Place).NotEmpty();
+        }
+    }
+}
