+++ TestApp/Validators/UpdateHeroModelValidation.cs
@@ -1,0 +1,14 @@
+﻿namespace TestApp.Validators
+{
+    public class UpdateHeroModelValidation : AbstractValidator<UpdateHeroModel>
+    {
+        public UpdateHeroModelValidation()
+        {
+            RuleFor(x => x).NotEmpty();
+            RuleFor(x => x.Id).NotEmpty();
+            RuleFor(x => x.Name).NotEmpty();
+            RuleFor(x => x.FirstName).NotEmpty();
+            RuleFor(x => x.Place).NotEmpty();
+        }
+    }
+}
